pyproject.toml
[tool.poetry]
name = "drf-referral"
version = "0.1.0"
description = ""
authors = ["Asadhimself <theavenger1712@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.4"
djangorestframework = "^3.15.1"
drf-spectacular = "^0.27.2"
django-phonenumber-field = "^7.3.0"
phonenumbers = "^8.13.35"
psycopg2 = "^2.9.9"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
